we are somewhat-officially misra-compliant

+++ cruise-control/scs-impl.h
@@ -1,11 +1,11 @@
 #ifndef CRUISE_CONTROL_IMPL_H
 #define CRUISE_CONTROL_IMPL_H
 
-#define lever5_auto_time 2000
-#define lever5_frequency 1000
+#define lever5_auto_time ((size_t) 2000)
+#define lever5_frequency ((size_t) 1000)
 
-#define lever7_auto_time 2000
-#define lever7_frequency 2000
+#define lever7_auto_time ((size_t) 2000)
+#define lever7_frequency ((size_t) 2000)
 
 void scs_do_step(void);
